# my-blog-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  my-blog-backend_internal_api_v1_dto_request.CreateUserRequest:
    properties:
      email:
        type: string
      nickname:
        maxLength: 50
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  my-blog-backend_internal_api_v1_dto_request.UpdateUserRequest:
    properties:
      avatar:
        type: string
      bio:
        type: string
      email:
        type: string
      github:
        type: string
      nickname:
        maxLength: 50
        type: string
      password:
        minLength: 6
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
      website:
        type: string
    type: object
  my-blog-backend_internal_api_v1_dto_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@myblog.com
    name: API Support
    url: http://www.myblog.com/support
  description: This is a blog backend API with Swagger documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: My Blog API
  version: "1.0"
paths:
...
  /api/v1/stats/articles/hot:
    get:
      consumes:
      - application/json
      parameters:
      - default: 10
        description: 限制数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取热门文章
      tags:
      - 统计管理
  /api/v1/stats/articles/trend:
    get:
      consumes:
      - application/json
      parameters:
      - default: 7
        description: 天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取文章发布趋势
      tags:
      - 统计管理
...
  /api/v1/stats/tags/hot:
    get:
      consumes:
      - application/json
      parameters:
      - default: 10
        description: 限制数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取热门标签
      tags:
      - 统计管理
...
  /api/v1/stats/users/active:
    get:
      consumes:
      - application/json
      parameters:
      - default: 7
        description: 天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取活跃用户趋势
      tags:
      - 统计管理
...
  /api/v1/stats/users/growth:
    get:
      consumes:
      - application/json
      parameters:
      - default: 7
        description: 天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取用户增长趋势
      tags:
      - 统计管理
...
  /api/v1/stats/views/trend:
    get:
      consumes:
      - application/json
      parameters:
      - default: 7
        description: 天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取浏览量趋势
      tags:
      - 统计管理
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 创建用户
      tags:
      - 用户管理
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 更新用户
      tags:
      - 用户管理
  /api/v1/users/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 更新用户状态
      tags:
      - 用户管理
  /api/v1/users/current:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取当前登录用户信息
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 更新当前用户信息
      tags:
      - 用户管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"